requirements.txt
fastapi
pydantic
uvicorn
opencv-python-headless
facemorpher
python-multipart
cloudinary
python-dotenv